09/02/2026 - Migrating Bot from Discord.js v12 to v14 and updated outdated dependencies - Started work on bringing this project back to life

+++ commands/fun/webhook.js
@@ -4,12 +4,16 @@ module.exports = {
     usage: '<user> <message>',
     category: 'fun',
     run: async (client, message, args) => {
+        const { PermissionFlagsBits } = require('discord.js');
+        if (!message.guild.members.me.permissions.has(PermissionFlagsBits.ManageWebhooks)) {
+            return message.channel.send("I don't have the `Manage Webhooks` permission to do this!");
+        }
         let user = message.mentions.members.first() || message.guild.members.cache.get(args[0])
-        if(!user) return message.channel.send('Please provide a user!')
-        const webhook = await message.channel.createWebhook(user.displayName, {
-            avatar: user.user.displayAvatarURL(),
-            channel: message.channel.id
-        })
+        if (!user) return message.channel.send('Please provide a user!')
+        const webhook = await message.channel.createWebhook({
+            name: user.displayName,
+            avatar: user.displayAvatarURL({ dynamic: true })
+        });
         await webhook.send(args.slice(1).join(' ')).then(() => {
             webhook.delete()
         })
